Moved things around; js_only.
