+++ assets/src/application/ui/input/confirm/ConfirmDelete.tsx
